api_example/requirements.txt
Django==2.2.12
gunicorn==20.0.4
djangorestframework